Addressing inability to display TIF in browser for #4
The solution is to convert to jpeg format when processing tif
images for thumbnails, and to link to full-size jpeg image data
instead of the file itself from the clickable links.

This should work, but for another problem which is that PIL's
conversion seems to be messing with the colours of the tif images.

But it's a start :)